back to original 2D

+++ Ares/src/Ares/Core/Time.h
@@ -14,7 +14,7 @@ namespace Ares
 		const static double GetTime() { return m_Time; }
 		const static double GetDeltaTime() { return m_DeltaTime; }
 		const static long long unsigned int GetFrame() { return m_Frame; }
-		const static uint32_t GetFPS() { return 1.0f / m_DeltaTime; }
+		const static uint32_t GetFPS() { return (uint32_t)(1.0f / m_DeltaTime); }
 	private:
 		static inline double m_Time = 0.0;
 		static inline double m_DeltaTime = 0.0;
